Bug 1745912 - Use the same type in IDL and implementation file for MediaTransportHandler. r=pehrsons
Depends on D133733

Differential Revision: https://phabricator.services.mozilla.com/D133737

+++ dom/media/webrtc/jsapi/MediaTransportHandlerIPC.cpp
@@ -179,7 +179,7 @@ void MediaTransportHandlerIPC::SetProxyConfig(
 
 void MediaTransportHandlerIPC::EnsureProvisionalTransport(
     const std::string& aTransportId, const std::string& aLocalUfrag,
-    const std::string& aLocalPwd, size_t aComponentCount) {
+    const std::string& aLocalPwd, int aComponentCount) {
   mInitPromise->Then(
       mCallbackThread, __func__,
       [=, self = RefPtr<MediaTransportHandlerIPC>(this)](bool /*dummy*/) {
